+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 - Make checkbox column a bit wider for multiselects.
 ### Fixed
 - Multiselect conditions code had a couple of problems.
+- Fix to conditions-handling for listing data.
 
 ## [0.1.8] - 2018-08-31
 ### Added